Silence some warnings.

+++ SBCenteringClipView.m
@@ -165,7 +165,7 @@
 	[self centerDocument];
 }
 
--(void)setFrameRotation:(float)angle
+-(void)setFrameRotation:(CGFloat)angle
 {
 	[super setFrameRotation:angle];
 	//[self centerDocument];
